修正: クラスの役割りを整理
